refactor: combine type definitions into one file
