+++ packages/core/src/index.ts
@@ -5,3 +5,4 @@ export * from './svg'
 export * from './throttle'
 export * from './shared'
 export * from './handler'
+export * from './types'
